requirements.txt
customtkinter
Pillow
pynput
pyperclip
pyautogui
mss
PyInstaller